deploy/deploy_requirements.txt
onnxruntime
boto3
python-dotenv
numpy
Pillow